fstests-results: fix regression list and add dmesg list
Sometimes subsequent runs will pass, especially for things like lockdep
where it fails once and then is fine the next time.  If there are any
failures in the last week count it as a regression and carry on.  Add
another column for dmesg failures so we can find serious problems.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>